Move location tracking out of the driver (#6)
This now let's us use the @n location actions in the parser to
get the location of each token.

Bison calls yylex with a pointer to a location an expects the
scanner to update the location. Before the this location was
ignored and a sperate location in the driver was used. However,
bison wasn't aware of that and therefore the location actions
pointed to wrong location.

More about bisons location action:
https://www.gnu.org/software/bison/manual/html_node/Actions-and-Locations.html